moved helpers to SentimentModel
